+++ atlasdb-cassandra-multinode-tests/src/test/java/com/palantir/cassandra/multinode/CassandraSchemaLockTest.java
@@ -21,7 +21,7 @@ import com.google.common.base.Throwables;
 import com.palantir.atlasdb.AtlasDbConstants;
 import com.palantir.atlasdb.cassandra.CassandraKeyValueServiceConfig;
 import com.palantir.atlasdb.cassandra.CassandraKeyValueServiceRuntimeConfig;
-import com.palantir.atlasdb.containers.ContainersV2;
+import com.palantir.atlasdb.containers.Containers;
 import com.palantir.atlasdb.containers.ThreeNodeCassandraCluster;
 import com.palantir.atlasdb.keyvalue.api.TableReference;
 import com.palantir.atlasdb.keyvalue.cassandra.CassandraKeyValueService;
@@ -51,8 +51,8 @@ public class CassandraSchemaLockTest {
     private static final int THREAD_COUNT = 4;
 
     @RegisterExtension
-    public static final ContainersV2 CONTAINERS =
-            new ContainersV2(CassandraSchemaLockTest.class).with(new ThreeNodeCassandraCluster());
+    public static final Containers CONTAINERS =
+            new Containers(CassandraSchemaLockTest.class).with(new ThreeNodeCassandraCluster());
 
     private final ExecutorService executorService = Executors.newFixedThreadPool(THREAD_COUNT);
