+++ atlasdb-container-test-utils/src/test/java/com/palantir/atlasdb/containers/ContainerTest.java
@@ -17,28 +17,18 @@ package com.palantir.atlasdb.containers;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import java.util.Arrays;
-import java.util.Collection;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
+import java.util.List;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.MethodSource;
 
-/* TODO(boyoruk): Delete when JUnit5 upgrade is over. */
-@RunWith(Parameterized.class)
 public class ContainerTest {
-    @Parameterized.Parameters(name = "With container {0}")
-    public static Collection<Object[]> parameters() {
-        return Arrays.asList(new Object[][] {{new CassandraContainer()}});
+    public static List<Container> containers() {
+        return List.of(new CassandraContainer(), new ThreeNodeCassandraCluster());
     }
 
-    private final Container container;
-
-    public ContainerTest(Container container) {
-        this.container = container;
-    }
-
-    @Test
-    public void dockerComposeFileShouldExist() {
+    @ParameterizedTest
+    @MethodSource("containers")
+    public void dockerComposeFileShouldExist(Container container) {
         assertThat(ContainerTest.class.getResource(container.getDockerComposeFile()))
                 .isNotNull();
     }
